Rework pathfinding
The current direction (and count of steps in that direction) must be
part of the state, because once we land on a place, we can't assume
that the only way to get to that place is the way we came. Makes the
pathfinding much more complex (the graph is *much* larger), but it
gives us the right answer, so we'll allow it.